feat: season 타입별 컴포넌트 분기
