+++ package.json
@@ -10,8 +10,10 @@
     "url": "https://github.com/Awesome-Sprinters/react-season-component"
   },
   "dependencies": {
+    "gsap": "^3.9.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-gsap": "^3.2.1",
     "react-scripts": "5.0.0",
     "styled-components": "^5.3.3",
     "typescript": "^4.5.5"
