+++ src/lib/index.tsx
@@ -1,3 +1,3 @@
-export { default as Season } from './Season';
-export { default as Time } from './Time';
-export { default as Weather } from './Weather';
+export { default as Season } from "./Season";
+// export { default as Time } from './Time';
+export { default as Weather } from "./Weather";
